Add multi-platform support (#2)
Add support for multi-platform images. Testing the type of the image is done through inspecting its manifest. When it is a single-platform image, it is pulled, retagged and pushed, as before. When it is a multi-platform image, the content of the manifest is further analysed, and for each platform that it contains: the image for that platform will be pulled, retagged (with a tag containing the target platform) and pushed. At the same time, a manifest with the target tag is constructed and amended, with every platform-specific tag that was just pushed.